+++ Logic_ex.c
@@ -1,6 +1,6 @@
 /*!\file Logic_ex.c
 ** \author SMFSW
-** \copyright MIT (c) 2017-2022, SMFSW
+** \copyright MIT (c) 2017-2023, SMFSW
 ** \brief Extensions for logic variables (R/W)
 **/
 /****************************************************************/