+++ Logic_in.c
@@ -1,6 +1,6 @@
 /*!\file Logic_in.c
 ** \author SMFSW
-** \copyright MIT (c) 2017-2022, SMFSW
+** \copyright MIT (c) 2017-2023, SMFSW
 ** \brief Logic input handling
 **/
 /****************************************************************/