+++ USB_CDC_IF_ex.c
@@ -1,6 +1,6 @@
 /*!\file USB_CDC_IF_ex.c
 ** \author SMFSW
-** \copyright MIT (c) 2017-2022, SMFSW
+** \copyright MIT (c) 2017-2023, SMFSW
 ** \brief Extensions for USB VCP
 **
 ** \code